requirements.txt
cffi==1.16.0
cryptography==41.0.7
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
